:recycle: Rename the file to better reflect its use
